% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_specs}
\alias{compare_specs}
\title{Compare original vs minimal adjustment specification}
\usage{
compare_specs(
  dag,
  formula,
  data,
  exposure,
  outcome,
  engine = stats::lm,
  engine_args = list()
)
}
\arguments{
\item{dag}{A \code{dagitty} DAG.}

\item{formula}{A model formula (e.g., \code{Y ~ X + Z}).}

\item{data}{A data.frame with the variables used in \code{formula}.}

\item{exposure}{Character; exposure variable name.}

\item{outcome}{Character; outcome variable name.}

\item{engine}{A fitting function (default \code{stats::lm}).}

\item{engine_args}{List of extra arguments passed to \code{engine}.
NOTE: will want to edit so it works with fixed/random effects and other
common uses. eventually will want to make sure it works with everything}
}
\value{
An object of class \code{DAGassist_compare} with elements:
\code{validation}, \code{original}, \code{minimal}, \code{controls}, \code{formulas}.
}
\description{
Fits the user-supplied model and a minimal-adjustment model (via
\code{dagitty::adjustmentSets()}), after validating inputs.
}
\examples{
# compare_specs(test_confounder, Y ~ X + Z, test_df, "X", "Y")
}
