% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_helpers.R
\name{glance_custom.fixest}
\alias{glance_custom.fixest}
\title{Custom GOF fields for fixest models in modelsummary tables}
\usage{
\method{glance_custom}{fixest}(x, ...)
}
\description{
This is a method for modelsummary::glance_custom().
}
\keyword{internal}
