% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{partial_test_df}
\alias{partial_test_df}
\title{Partial demo data (missing some variables)}
\format{
A data frame with 3 variables: \code{X}, \code{Y}, \code{Z}.
}
\source{
Simulated for examples in the package.
}
\usage{
partial_test_df
}
\description{
Partial demo data (missing some variables)
}
\keyword{datasets}
