% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{pick_minimal_controls}
\alias{pick_minimal_controls}
\title{compute minimal adjustment sets and pick one deterministically}
\usage{
pick_minimal_controls(dag, exposure, outcome)
}
\arguments{
\item{dag}{A \code{dagitty} DAG object.}

\item{exposure}{Character; exposure node name (X).}

\item{outcome}{Character; outcome node name (Y).}
}
\value{
sorted character vector with the minimal adjustment set
}
\description{
compute minimal adjustment sets and pick one deterministically
}
\details{
This is preferable to \code{dagitty::adjustmentSets()} because it picks a single
minimal adjustment set automatically and without a bunch of arguments. will
need to adapt this to deal with multiple adjustment sets and canonical
adjustment sets eventually.
}
\examples{
\dontshow{if (requireNamespace("dagitty", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
d <- ggdag::dagify(
Y ~ X + M + Z,
M ~ X + Z,
C ~ X + Y,
exposure = "X",
outcome = "Y")
# M: mediator / Z: confounder / C: collider

# identify the minimal adjustment set
pick_minimal_controls(d, exposure = "X", outcome = "Y")
# expected: "Z"
\dontshow{\}) # examplesIf}
}
