% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.DAGassist_compare}
\alias{print.DAGassist_compare}
\title{Pretty print compare_specs()
NOTE: may want to clean for kable or latex output later}
\usage{
\method{print}{DAGassist_compare}(x, ...)
}
\description{
Pretty print compare_specs()
NOTE: may want to clean for kable or latex output later
}
