% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{print.DAGassist_roles}
\alias{print.DAGassist_roles}
\title{Print node classifications (aligned)}
\usage{
\method{print}{DAGassist_roles}(x, n = Inf, ...)
}
\arguments{
\item{x}{Output of classify_nodes() (class "DAGassist_roles")}

\item{n}{Max rows to print (default all)}

\item{...}{(ignored)}
}
\value{
Invisibly returns x
}
\description{
Print node classifications (aligned)
}
