% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print.DAGassist_validation}
\alias{print.DAGassist_validation}
\title{Minimal, clean printout for validation results with color coding}
\usage{
\method{print}{DAGassist_validation}(x, n = 10, ...)
}
\arguments{
\item{x}{the list (class \code{out}) from validate_spec}

\item{n}{Max number of issues to show (default 10).}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Minimal, clean printout for validation results with color coding
}
