% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_collider}
\alias{test_collider}
\title{Demo DAG with a collider}
\format{
A \code{dagitty} object with nodes: X,Y,C
}
\source{
Simulated for examples in the package.
}
\usage{
test_collider
}
\description{
Demo DAG with a collider
}
\keyword{datasets}
