% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_complex}
\alias{test_complex}
\title{Demo DAG with confounder, mediator, collider}
\format{
A \code{dagitty} object with nodes: X,Y,Z,M,C,A
}
\source{
Simulated for examples in the package.
}
\usage{
test_complex
}
\description{
Demo DAG with confounder, mediator, collider
}
\keyword{datasets}
