% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_cycle}
\alias{test_cycle}
\title{Cyclic DAG used to trigger acyclicity errors}
\format{
A \code{dagitty} object with variables X,Y,Z
}
\source{
Simulated for examples in the package.
}
\usage{
test_cycle
}
\description{
Cyclic DAG used to trigger acyclicity errors
}
\keyword{datasets}
