% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_dags}
\alias{test_dags}
\alias{test_confounder}
\alias{test_collider}
\alias{test_mediator}
\title{Example DAGs for demos and tests}
\format{
Each object is a \code{dagitty} DAG.

An object of class \code{dagitty} of length 1.

An object of class \code{dagitty} of length 1.

An object of class \code{dagitty} of length 1.
}
\usage{
test_confounder

test_collider

test_mediator
}
\description{
Three tiny DAGs: a confounder case, a collider, and a mediator.
They are stored as \code{dagitty} DAG objects.
}
\details{
\itemize{
\item \code{test_confounder}: U -> X, U -> Y, X -> Y
\item \code{test_collider}:  X -> C <- Y
\item \code{test_mediator}:  X -> M -> Y
}
}
\examples{
data(test_confounder, package = "DAGassist")
test_confounder

}
\keyword{datasets}
