% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{Demo data frame consistent with demo DAG}
\format{
A data frame with 400 rows and 6 variables: \code{X}, \code{Y}, \code{Z}, \code{C}, \code{M}, \code{A}.
}
\source{
Simulated for examples in the package.
}
\usage{
test_df
}
\description{
Demo data frame consistent with demo DAG
}
\keyword{datasets}
