% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_mediator}
\alias{test_mediator}
\title{Demo DAG with a confounder}
\format{
A \code{dagitty} object with nodes: X,Y,M
}
\source{
Simulated for examples in the package.
}
\usage{
test_mediator
}
\description{
Demo DAG with a confounder
}
\keyword{datasets}
