% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_multimin}
\alias{test_multimin}
\title{DAG with two minimal adjustment sets}
\format{
A \code{dagitty} object with variables M,W,X,Y,Z
}
\source{
Simulated for examples in the package
}
\usage{
test_multimin
}
\description{
DAG with two minimal adjustment sets
}
\keyword{datasets}
