% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dags.R
\docType{data}
\name{test_multimin4}
\alias{test_multimin4}
\title{DAG with four minimal adjustment sets}
\format{
A \code{dagitty} object with variables M,N,W,X,Y,Z
}
\source{
Simulated for examples in the package.
}
\usage{
test_multimin4
}
\description{
DAG with four minimal adjustment sets
}
\keyword{datasets}
