% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_seqg.R
\name{tidy.seqg}
\alias{tidy.seqg}
\title{Tidy a DirectEffects sequential_g model (class 'seqg')}
\usage{
\method{tidy}{seqg}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A seqg object from DirectEffects::sequential_g()}

\item{conf.int}{Logical; include confidence intervals}

\item{conf.level}{Confidence level}

\item{...}{Unused}
}
\description{
Tidy a DirectEffects sequential_g model (class 'seqg')
}
