% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{bad_controls_in}
\alias{bad_controls_in}
\title{flag bad controls (mediator/collider/desc of Y) among a candidate set}
\usage{
bad_controls_in(dag, controls, exposure, outcome)
}
\arguments{
\item{dag}{A \code{dagitty} DAG object.}

\item{controls}{Character vector of variable names.}

\item{exposure}{Character; exposure node name (X).}

\item{outcome}{Character; outcome node name (Y).}
}
\value{
A character vector (possibly empty) containing the elements of
\code{controls} that are identified as "bad controls".

This is essentially the inverse of \code{pick_minimal_controls()}, as it returns
bad controls, rather than the minimal/canonical set of good controls
}
\description{
flag bad controls (mediator/collider/desc of Y) among a candidate set
}
\examples{
# Which variables in a formula are bad controls, given the DAG?
bad_controls_in(toy_dag, controls = c("M", "C", "Z"))
}
