% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{bad_controls_in}
\alias{bad_controls_in}
\title{List "bad controls" among a candidate set
NOTE: will flag mediator / collider / descendant}
\usage{
bad_controls_in(dag, controls, exposure, outcome)
}
\description{
List "bad controls" among a candidate set
NOTE: will flag mediator / collider / descendant
}
