% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{dag_assist}
\alias{dag_assist}
\title{need to populate this roxy with info later}
\usage{
dag_assist(
  dag,
  formula,
  data,
  exposure,
  outcome,
  engine = stats::lm,
  engine_args = list()
)
}
\description{
need to populate this roxy with info later
}
