% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{DAGassist}
\alias{DAGassist}
\title{Main function
Run the DAGassist pipeline and print a compact report, tying it all together}
\usage{
DAGassist(
  dag,
  formula,
  data,
  exposure,
  outcome,
  engine = stats::lm,
  engine_args = list(),
  verbose = TRUE,
  type = c("console", "latex", "word", "docx", "excel", "xlsx", "text", "txt"),
  out = NULL,
  imply = FALSE
)
}
\arguments{
\item{dag}{a valid dagitty object}

\item{formula}{a regression formula \emph{or} a single engine call--e.g.
\code{feols(Y~X+Z | A, data = df, cluster = ~id)}}

\item{data}{a valid data frame, (optional if supplied via formula call)}

\item{exposure}{optional; inferred from the DAG if missing}

\item{outcome}{optional; inferred from DAG if missing}

\item{engine}{the regression engine; default \code{stats::lm}}

\item{engine_args}{a list with exra engine args---merged with any found in call}

\item{verbose}{default TRUE-- TRUE/FALSE -- suppresses formulas and notes.}

\item{type}{Output type: "console" (default) or "latex" or "word"}

\item{out}{File path for latex fragment or word doc, default NULL}

\item{imply}{logical. If TRUE, DAGassist will add variables not included in
the user's main formula, based on the relationships in the DAG. If FALSE (default),
only the original model is shown}
}
\value{
a list of class \code{out} with the \code{DAGassist_report}, which has values
validation: runs \code{validate_spec} from validate.R
roles: runs \code{classify_nodes} from classify.R
bad_in_user: does the user include mediator, collider or descendant as controls?
controls_minimal: runs \code{pick_minimal_controls} from compare.R
formulas: a list with the original regression formula + minimal formula
models: a list with the full original regression call + minimal regression call
}
\description{
Main function
Run the DAGassist pipeline and print a compact report, tying it all together
}
\examples{
\dontshow{if (requireNamespace("dagitty", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Use package datasets to avoid re-simulating
data(test_df, package = "DAGassist")
data(test_complex, package = "DAGassist")
DAGassist(test_complex, Y ~ X + Z + C + M, test_df, exposure = "X", outcome = "Y")
\dontshow{\}) # examplesIf}
}
