% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{pick_minimal_controls}
\alias{pick_minimal_controls}
\title{Pick one minimal adjustment set deterministically}
\usage{
pick_minimal_controls(dag, exposure, outcome)
}
\description{
Pick one minimal adjustment set deterministically
}
