% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{print.DAGassist_report}
\alias{print.DAGassist_report}
\title{print the DAGassist_report}
\usage{
\method{print}{DAGassist_report}(x, ...)
}
\arguments{
\item{x}{Output of DAGassist() (class "report")}

\item{...}{(ignored)}
}
\value{
Invisibly returns x
}
\description{
print the DAGassist_report
}
\examples{
\dontshow{if (requireNamespace("dagitty", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Use package datasets to avoid re-simulating
data(test_df, package = "DAGassist")
data(test_complex, package = "DAGassist")
DAGassist(test_complex, Y ~ X + Z + C + M, test_df, exposure = "X", outcome = "Y")
\dontshow{\}) # examplesIf}
}
