% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_spec}
\alias{validate_spec}
\title{Validate dag, formula, data before fitting and ensure exposure and outcome are
specified}
\usage{
validate_spec(dag, formula, data, exposure, outcome)
}
\arguments{
\item{dag}{A \code{dagitty} DAG object.}

\item{formula}{a regression call}

\item{data}{data frame}

\item{exposure}{Character; exposure node name (X).}

\item{outcome}{Character; outcome node name (Y).}
}
\value{
A list of class \code{DAGassist_validation} with values:
\code{ok} (logical), \code{issues} (data.frame), \code{vars} (list)
}
\description{
Validate dag, formula, data before fitting and ensure exposure and outcome are
specified
}
