% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_spec}
\alias{validate_spec}
\alias{print.DAGassist_validation}
\title{Validate dag, formula, data before fitting and ensure exposure and outcome are
specified}
\usage{
validate_spec(dag, formula, data, exposure, outcome)

\method{print}{DAGassist_validation}(x, n = 10, ...)
}
\arguments{
\item{dag}{}

\item{formula}{}

\item{data}{}

\item{exposure}{}

\item{outcome}{}

\item{n}{Max number of issues to show (default 10).}
}
\value{
A list of class \code{DAGassist_validation} with values:
\code{ok} (logical), \code{issues} (data.frame), \code{vars} (list)

Invisibly returns \code{x}.
}
\description{
Validate dag, formula, data before fitting and ensure exposure and outcome are
specified

Minimal, clean printout for validation results with color coding
}
\examples{
v <- validate_spec(dag, Y ~ X + Z, df, exposure = "X", outcome = "Y")
print(v)
}
